% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_cores.R
\name{available_cores}
\alias{available_cores}
\title{Get The Number of Parallel Cores}
\usage{
available_cores()
}
\value{
The number of cores allowed on the current machine.
}
\description{
Get The Number of Parallel Cores
}
