% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc_releases.R
\name{bioc_releases}
\alias{bioc_releases}
\title{Gets the Version and Release Date for All Bioconductor Releases}
\usage{
bioc_releases()
}
\value{
A data.frame.
}
\description{
Gets the Version and Release Date for All Bioconductor Releases
}
