% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Run Reverse Dependency Checks}
\usage{
check(bioc = TRUE, timeout = 60)
}
\arguments{
\item{bioc}{(logical) Whether or not reverse dependencies
on Bioconductor should be checked.}

\item{timeout}{(numeric or \link[base:difftime]{base::difftime}) The maximum
run-time (in minutes) of 'R CMD check' for one package and
version, where "old" and "new" counts as two seperate versions.}
}
\value{
What \code{\link[revdepcheck:revdep_check]{revdepcheck::revdep_check()}} returns.
}
\description{
Run Reverse Dependency Checks
}
\details{
This function does:
\enumerate{
\item Gives a note if a \code{R_CHECK_ENVIRON} file exists
\item Lists any \verb{R_CHECK_*} and \verb{_R_CHECK_*} variable sets
\item Calls \code{precheck()} if such a function is availble
\item Calls \code{\link[revdepcheck:revdep_check]{revdepcheck::revdep_check()}}
}
}
