% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crancache_cleanup.R
\name{crancache_cleanup}
\alias{crancache_cleanup}
\title{Clean up the local crancache}
\usage{
crancache_cleanup(dryrun = FALSE)
}
\arguments{
\item{dryrun}{If TRUE, the cache is not cleaned up.}
}
\value{
Invisibly returns currently cached packages.
}
\description{
Looks for packages in the local crancache that are no longer available
online. For example, this removes recently archive CRAN packages and
Bioconductor package no longer available.
}
