% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_crancache.R
\name{is_in_crancache}
\alias{is_in_crancache}
\title{Checks whether or not a package is already in the binary crancache}
\usage{
is_in_crancache(pkgs)
}
\arguments{
\item{pkgs}{(character vector) Zero or more package names}
}
\value{
A named logical vector indicating whether or not
each package is in the binary crancache.
}
\description{
Checks whether or not a package is already in the binary crancache
}
