% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_cores.R
\name{num_workers}
\alias{num_workers}
\title{Get The Number of Concurrent Check Workers}
\usage{
num_workers()
}
\value{
The number of concurrent workers allowed on the current machine.
}
\description{
Get The Number of Concurrent Check Workers
}
