% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pkgs.R
\name{parse_pkgs}
\alias{parse_pkgs}
\title{Extract packages names from messy strings}
\usage{
parse_pkgs(pkgs)
}
\arguments{
\item{pkgs}{(character vector) Extract package names
from one or more strings with our without commas, spaces,
quotation marks, etc.}
}
\value{
(character vector) A sorted set of unique package
names.
}
\description{
Extract packages names from messy strings
}
