% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_packages.R
\name{required_packages}
\alias{required_packages}
\title{Find all packages required by a specific package}
\usage{
required_packages(packages, ...)
}
\description{
Find all packages required by a specific package
}
