% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_init.R
\name{revdep_init}
\alias{revdep_init}
\title{Initiate the revdepcheck database}
\usage{
revdep_init(pkg = ".")
}
\arguments{
\item{pkg}{(character) Path to package.}
}
\value{
Nothing.
}
\description{
Initiate the revdepcheck database
}
