% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_over_time.R
\name{revdep_over_time}
\alias{revdep_over_time}
\alias{cran_revdep_on_date}
\title{Get the Number of Reverse Dependencies on CRAN}
\usage{
revdep_over_time(pkgs, dates, none = NA_integer_, force = FALSE)

cran_revdep_on_date(package, date, force = FALSE)
}
\arguments{
\item{pkgs}{(character vector) One or more packages.}

\item{dates}{(POSIXct vector) One or more dates.}

\item{none}{(integer) The default value if there are no dependencies.}

\item{force}{(logical) If TRUE, cached values are ignored.}

\item{package}{(character) A package name.}

\item{date}{(POSIXct) A date.}
}
\value{
A data.frame
}
\description{
Get the Number of Reverse Dependencies on CRAN
}
\details{
This function uses the \href{https://github.com/r-hub/evercran}{evercran}
(\code{"EverCRAN"}) to identify reverse package dependencies at particular
dates.
It also supports the
\href{https://packagemanager.posit.co/}{Posit Public Package Manager}
(\code{"PPPM"}), which goes back to October 2017, as an alternative.
The default is \code{EverCRAN}, which can be overridden by R option
\code{revdepcheck.extras.snapshot.source}, which in turn can be set by
environment variable \code{R_REVDEPCHECK_EXTRAS_SNAPSHOT_SOURCE} on
package load.
}
