% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_pkgs_with_status.R
\name{revdep_pkgs_with_status}
\alias{revdep_pkgs_with_status}
\title{Get packages with a certain check status}
\usage{
revdep_pkgs_with_status(pkg = ".", status = c("error", "failure"))
}
\arguments{
\item{pkg}{(character) Path to package.}

\item{status}{(character vector) A status to filter checked
packages by.
If \code{"error"}, then packages for which the 'new' version failed are returned.
If \code{"failure"}, then packages with an internal "E" status are returned.}
}
\value{
(character vector) Zero or more package names.
}
\description{
Get packages with a certain check status
}
