% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_precache.R
\name{revdep_precache}
\alias{revdep_precache}
\title{Precache package installs required for reverse-dependencies package checks}
\usage{
revdep_precache(package, temp_lib_path = tempfile(pattern = "dir"), ...,
  dryrun = FALSE)
}
\arguments{
\item{package}{(character string) Name of package}

\item{temp_lib_path}{...}

\item{dryrun}{(logical) If TRUE, then no packages are cached.}

\item{\ldots}{Not used.}
}
\description{
Precache package installs required for reverse-dependencies package checks
}
