% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_preinstall.R
\name{revdep_preinstall}
\alias{revdep_preinstall}
\alias{revdep_preinstall_update}
\title{Pre-install and updated reverse-dependency packages}
\usage{
revdep_preinstall(
  pkgs,
  skip = TRUE,
  chunk_size = availableCores(),
  temp_lib_path = revdep_preinstall_libs()[1]
)

revdep_preinstall_update()
}
\arguments{
\item{pkgs}{(character vector) Packages to be pre-installed.}

\item{skip}{If TRUE, packages already in the binary crancache cache will
be skipped.  If FALSE, all packages will be pre-installed.}

\item{chunk_size}{If \code{1}, packages are installed one by one.
If \code{2}, packages are installed two by two, and so on.
If \code{Inf}, all packages are installed in one go.
This affects how often the \strong{crancache} package library cache is
updated. The smaller the chunk size is the more frequently it is
updated.}

\item{temp_lib_path}{(character string) The folder where to
install packages during pre-installation.}
}
\value{
Nothing.
}
\description{
Pre-install and updated reverse-dependency packages
}
\details{
Reverse-dependency packages are pre-installed to custom package
library folders with suffix \file{-revdepcheck} that lives next
to your default library folders.
Packages are forced to be installed from source
(https://github.com/r-lib/crancache/issues/46), regardless of
binary packages already existing in the cache or not.
}
