% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_readme_packages.R
\name{revdep_readme_packages}
\alias{revdep_readme_packages}
\title{Gets Packages Previously Checked}
\usage{
revdep_readme_packages(file = "revdep/README.md")
}
\arguments{
\item{file}{(character) Pathname to the \file{revdep/README.md}}
}
\value{
A data.frame with columns \code{package} and \code{version}.
}
\description{
Gets Packages Previously Checked
}
