% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_required_packages.R
\name{revdep_required_packages}
\alias{revdep_required_packages}
\title{Find all required reverse package dependencies}
\usage{
revdep_required_packages(package, ...)
}
\description{
Find all required reverse package dependencies
}
