% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_reset.R
\name{revdep_reset}
\alias{revdep_reset}
\title{Resets revdep check folders}
\usage{
revdep_reset(pkg = ".")
}
\arguments{
\item{pkg}{(character) Path to package.}
}
\value{
Nothing.
}
\description{
Resets revdep check folders
}
