% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_this_package.R
\name{revdep_this_package}
\alias{revdep_this_package}
\title{Get the name of package whose reverse dependencies should be checked}
\usage{
revdep_this_package()
}
\value{
(character string) The package name.
}
\description{
Get the name of package whose reverse dependencies should be checked
}
