% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run reverse-package dependency checks from the command line (CLI)}
\usage{
run(pkg = ".", ..., warn = 1L, args = base::commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{pkg}{Path to package or package name.  Default to the package
in the current working directory.}

\item{\dots}{Not used.}

\item{warn}{(integer) The warn level of \R option \code{warn} while running
checks.}

\item{args}{(character string) The command-line arguments.}
}
\value{
Nothing.
}
\description{
Run reverse-package dependency checks from the command line (CLI)
}
\section{Usage}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{Rscript -e revdepcheck.extras::run --args <options>

--help                Display this help page
--version             Display version

List packages:
--list-children       List reverse package dependencies
--list-grandchildren  List second-generation reverse package dependencies
--list-todo           List packages "added" that will be checked next run

Populate the 'crancache' database with installable package binaries:
--preinstall-todo     Pre-install all packages to be checked
--preinstall <pkgs>   Pre-install specified packages
--preinstall-children Install all reverse dependencies
--preinstall-update   Install packages that have been updated since last run
--preinstall-error    Install packages that gave an "error" during checks
--preinstall-failure  Install packages that failed to be checked

Add and remove packages to be checked:
--reset               Full reset to restart checks from scratch
--todo-reset          Empty set of packages to be checked
--add <pkgs>          Add one or more packages to be checked
--add-broken          Add "broken" packages to be rechecked
--add-error           Add "errored" but not "failed" packages to be rechecked
--add-failure         Add "failed" but not "errored" packages to be rechecked
--add-children        Add first-generation reverse package dependencies
--add-grandchildren   Add second-generation reverse package dependencies
--add-all             Add first and second-generation dependencies
--add-updated         Add packages that have been updated since last run
--add-new             Add new packages that did not exist at last run
--rm <pkgs>           Remove one or more packages to be checked

Miscellanous setup tweaks:
--init                Create an empty revdep data base
--use-tmpdir          Host revdep/\{library,check\} under $TMPDIR

Show results:
--show-check <pkg>    Display 'old' and 'new' check results during checks
--list-error          List all packages that "errored"
--list-failure        List all packages that "failed"
--list-updated        List packages that have been updated since last run
--list-new            List new packages that did not exist at last run
}\if{html}{\out{</div>}}
}

\section{Examples}{

Here are some examples how to call this function from the command-line:

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{Rscript -e revdepcheck.extras::run --args --help
Rscript -e revdepcheck.extras::run --args --version
Rscript -e revdepcheck.extras::run --args --reset
Rscript -e revdepcheck.extras::run --args --add some, pkg, another
Rscript -e revdepcheck.extras::run --args --list-todo
Rscript -e revdepcheck.extras::run --args --preinstall-children
Rscript -e revdepcheck.extras::run --args --preinstall-todo

R_REVDEPCHECK_NUM_WORKERS=1 R_REVDEPCHECK_TIMEOUT=60 Rscript -e revdepcheck.extras::run
TAR_SKIP_CLUSTERMQ=true xvfb-run Rscript -e revdepcheck.extras::run
}\if{html}{\out{</div>}}
}

