% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{todo}
\alias{todo}
\title{List the Reverse-Dependency Packages to be Checked}
\usage{
todo()
}
\value{
(character vector; invisible) The name of the
packages to be tested.
}
\description{
List the Reverse-Dependency Packages to be Checked
}
