% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_children.R
\name{revdep_children}
\alias{revdep_children}
\title{Gets reverse-dependency packages of a package}
\usage{
revdep_children(pkg = NULL)
}
\arguments{
\item{pkg}{(character) The package name whose reverse dependencies should
be identified and returned.  If NULL (default), then the current package
according to \code{\link[=revdep_this_package]{revdep_this_package()}} is used.}
}
\value{
(character vector) List of package names

Internally, memoization is used such that all succeeding
look-ups in the same R session for the same package will
return immediately.
}
\description{
Gets reverse-dependency packages of a package
}
