% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_precache.R
\name{revdep_precache}
\alias{revdep_precache}
\title{Precache package installs required for reverse-dependencies package checks}
\usage{
revdep_precache(
  package = ".",
  temp_lib_path = tempfile(pattern = "dir"),
  ...,
  dryrun = FALSE
)
}
\arguments{
\item{package}{(character string) Package name.}

\item{temp_lib_path}{(character string) The folder where to
install packages during the pre-cache installation.  This
defaults to a temporary folder.}

\item{dryrun}{(logical) If TRUE, then no packages are cached.}

\item{\ldots}{Not used.}
}
\value{
(character vector) The packages that \emph{failed} to install.
}
\description{
Precache package installs required for reverse-dependencies package checks
}
\details{
This function populates the \pkg{crancache} package cache
that holds package source tarballs as well as package binaries.
This is done by calling \code{\link[crancache:install_packages]{crancache::install_packages()}} on
all reverse package dependencies as given by
\code{\link[=revdep_required_packages]{revdep_required_packages()}}.
These packages are installed to a temporary folder to avoid
adding them to your default package library folders.
}
