% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_preinstall.R
\name{revdep_preinstall}
\alias{revdep_preinstall}
\alias{revdep_preinstall_update}
\title{Pre-install and updated reverse-dependency packages}
\usage{
revdep_preinstall(pkgs)

revdep_preinstall_update()
}
\arguments{
\item{pkgs}{(character vector) Packages to be pre-installed.}
}
\value{
Nothing.
}
\description{
Pre-install and updated reverse-dependency packages
}
\details{
Reverse-dependency packages are pre-installed to custom package
library folders with suffix \file{-revdepcheck} that lives next
to your default library folders.
}
