% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep_todo_reset.R
\name{revdep_todo_reset}
\alias{revdep_todo_reset}
\title{Resets the set of reverse-package dependencies to be checked}
\usage{
revdep_todo_reset(pkg = ".")
}
\arguments{
\item{pkg}{(character) Path to package.}
}
\value{
Nothing.
}
\description{
Resets the set of reverse-package dependencies to be checked
}
