% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run reverse-package dependency checks from the command line (CLI)}
\usage{
run(..., warn = 1L, args = base::commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{\dots}{Not used.}

\item{warn}{(integer) The warn level of \R option \code{warn} while running
checks.}

\item{args}{(character string) The command-line arguments.}
}
\description{
Run reverse-package dependency checks from the command line (CLI)
}
\section{Usage}{
\preformatted{Rscript -e revdepcheck.extras::run --args <options>

--help                Display this help page
--version             Display version

List packages:
--list-children       List reverse package dependencies
--list-grandchildren  List second-generation reverse package dependencies

Populate the 'crancache' database with installable package binaries:
--preinstall-todo     Pre-install all packages to be checked
--preinstall <pkgs>   Pre-install specified packages
--preinstall-update   Populate crancache database ...
--preinstall-children Populate crancache database ...
--preinstall-error    Populate crancache database ...
--preinstall-failure  Populate crancache database ...

Add and remove packages to be checked:
--reset               Full reset to restart checks from scratch
--todo-reset          Empty set of packages to be checked
--todo                List packages to be checked
--add <pkgs>          Add one or more packages to be checked
--add-broken          Add "broken" packages to be rechecked
--add-error           Add "errored" packages to be rechecked
--add-failure         Add "failed" packages to be rechecked
--add-children        Add first-generation reverse package dependencies
--add-grandchildren   Add second-generation reverse package dependencies
--add-all             Add first and second-generation dependencies
--rm <pkgs>           Remove one or more packages to be checked

Show results:
--show-check <pkg>    Display 'old' and 'new' check results during checks
--list-error          List all packages that "errored"
--list-failure        List all packages that "failed"
}
}

\section{Examples}{

Here are some examples how to call this function from the command-line:\if{html}{\out{<div class="sh">}}\preformatted{Rscript -e revdepcheck.extras::run --args --help
Rscript -e revdepcheck.extras::run --args --version
Rscript -e revdepcheck.extras::run --args --reset
Rscript -e revdepcheck.extras::run --args --add some, pkg, another
Rscript -e revdepcheck.extras::run --args --todo
Rscript -e revdepcheck.extras::run --args --preinstall-children
Rscript -e revdepcheck.extras::run --args --preinstall-todo
}\if{html}{\out{</div>}}
}

