% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run reverse-package dependency checks from the command line (CLI)}
\usage{
run(warn = 1L, args = base::commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{warn}{(integer) The warn level of \R option \code{warn} while running
checks.}

\item{args}{(character string) The command-line arguments.}
}
\description{
Run reverse-package dependency checks from the command line (CLI)
}
\section{Examples}{

Here are some examples how to call this function from the command-line:\if{html}{\out{<div class="sh">}}\preformatted{Rscript -e revdepcheck.extras::run
Rscript -e revdepcheck.extras::run --reset
Rscript -e revdepcheck.extras::run --add some, pkg, another
Rscript -e revdepcheck.extras::run --todo
Rscript -e revdepcheck.extras::run --preinstall-children
Rscript -e revdepcheck.extras::run --preinstall-todo
}\if{html}{\out{</div>}}
}

