% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{todo}
\alias{todo}
\title{Get all the reverse-dependency packages to be checked}
\usage{
todo(pkg = ".", print = TRUE)
}
\arguments{
\item{pkg}{(character) Path to package.}

\item{print}{If TRUE, the list is printed, otherwise not.}
}
\value{
(character vector; invisible) The name of the
packages to be tested.
}
\description{
Get all the reverse-dependency packages to be checked
}
